lnet: remove locking protection ln_testproto_compat
lnet_net_lock(LNET_LOCK_EX) is a heavy-weight lock that is not
necessary here.  The bits in this field are only set rarely - via an
ioctl - and the pattern for reading and clearing them exactly matches
test_and_clear_bit().
So change the field to "unsigned long" (so test_and_clear_bit() can be
used), and use test_and_clear_bit(), discarding all other locking.

+++ include/linux/lnet/lib-types.h
@@ -1007,7 +1007,7 @@ struct lnet {
 	struct list_head		ln_lnds;
 
 	/* test protocol compatibility flags */
-	int				ln_testprotocompat;
+	unsigned long			ln_testprotocompat;
 
 	/*
 	 * 0 - load the NIs from the mod params
